Move AchievementTypeViewModel to ViewModels namespace
